Support django 2.0
On models with ForeignKey the on_delete parameter is a requirement now.
The default is models.CASCADE, so that is now being used.

The migrations are also changed because they now require it too. Only
the existing one has been modified, because otherwise django would
complain. The default on_delete is used anyway, so it shouldn't matter.